Bugfix : don't break with splits
Stupid bug, was iterating over the windows instead of the tabs to
generate the list of paths, which resulted in mismatches in the tabline.

Did not notice, because if there is only one win per tab it's
equivalent, and apparently I did not test with splits when developing
the unambiguity feature.

Fixed by actually iterating on the right thing.