actually fix colors
So I got it totally backwards in the previous commit.

Apparently, this is an issue with termguicolor not set. Not fully sure
however, cause I don't really get this highlight system. So I just make
it so that at least it doesn't crash and print a "warning" for future
self.
If anybody knows better, please let me know.